requirements.txt
pydantic~=2.4.2
httpx~=0.25.0
motor~=3.3.1
python-dotenv~=1.0.0